Restructure  directory
