+++ app.rb
@@ -1,0 +1,10 @@
+require './lib/person'
+require './lib/decorators/capitalize_deco'
+require './lib/decorators/trimmer_deco'
+
+person = Person.new(22, 'maximilianus')
+person.correct_name
+capitalized_person = CapitalizeDecorator.new(person)
+puts capitalized_person.correct_name
+capitalized_trimmed_person = TrimmerDecorator.new(capitalized_person)
+puts capitalized_trimmed_person.correct_name
