+++ lib/person.rb
@@ -26,11 +26,3 @@ class Person < Nameable
     age >= 18
   end
 end
-
-person = Person.new(22, 'maximilianus')
-person.correct_name
-
-capitalized_person = CapitalizeDecorator.new(person)
-p capitalized_person.correct_name
-capitalized_trimmed_person = TrimmerDecorator.new(capitalized_person)
-p capitalized_trimmed_person.correct_name
